+++ spec/integrations/pro/consumption/scheduling/with_single_thread_materialization_scheduler_spec.rb
@@ -5,34 +5,27 @@
 
 become_pro!
 
-class OneThreadScheduler < ::Karafka::Pro::Processing::Scheduler
+class OneThreadScheduler < ::Karafka::Pro::Processing::Schedulers::Base
   def initialize(queue)
     super
-    @mutex = Mutex.new
     @jobs_buffer = []
   end
 
   def schedule_consumption(jobs_array)
-    @mutex.synchronize do
-      jobs_array.each do |job|
-        @jobs_buffer << job
-        @queue.lock(job)
-      end
-
-      internal_manage
+    jobs_array.each do |job|
+      @jobs_buffer << job
+      @queue.lock(job)
     end
+
+    internal_manage
   end
 
   def manage
-    @mutex.synchronize do
-      internal_manage
-    end
+    internal_manage
   end
 
   def clear(group_id)
-    @mutex.synchronize do
-      @jobs_buffer.delete_if { |job| job.group_id == group_id }
-    end
+    @jobs_buffer.delete_if { |job| job.group_id == group_id }
   end
 
   private