+++ spec/integrations/pro/licensing/configure_with_valid_token_spec.rb
@@ -22,7 +22,7 @@ assert const_visible?('Karafka::Pro::Processing::Coordinator')
 assert const_visible?('Karafka::Pro::Processing::Partitioner')
 assert const_visible?('Karafka::BaseConsumer')
 assert const_visible?('Karafka::Pro::Processing::JobsBuilder')
-assert const_visible?('Karafka::Pro::Processing::Scheduler')
+assert const_visible?('Karafka::Pro::Processing::Schedulers::Default')
 assert const_visible?('Karafka::Pro::Routing::Features::LongRunningJob::Topic')
 assert const_visible?('Karafka::Pro::Routing::Features::LongRunningJob::Contracts')
 assert const_visible?('Karafka::Pro::Routing::Features::LongRunningJob::Config')
@@ -37,7 +37,7 @@ assert const_visible?('Karafka::Pro::Instrumentation::PerformanceTracker')
 assert_equal pro::Processing::StrategySelector, config.processing.strategy_selector.class
 assert_equal pro::Processing::Partitioner, config.processing.partitioner_class
 assert_equal pro::Processing::Coordinator, config.processing.coordinator_class
-assert_equal pro::Processing::Scheduler, config.processing.scheduler_class
+assert_equal pro::Processing::Schedulers::Default, config.processing.scheduler_class
 assert_equal pro::Processing::JobsQueue, config.processing.jobs_queue_class
 assert_equal pro::Processing::JobsBuilder, config.processing.jobs_builder.class
 assert_equal pro::ActiveJob::Dispatcher, config.active_job.dispatcher.class