+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.2.14 (Unreleased)
 - **[Feature]** Provide `Karafka::Admin#delete_consumer_group` and `Karafka::Admin#seek_consumer_group`.
 - **[Feature]** Provide `Karafka::App.assignments` that will return real-time assignments tracking.
+- [Enhancement] Make sure that the Scheduling API is thread-safe by default and allow for lock-less schedulers when schedulers are stateless.
 - [Enhancement] "Blockless" topics with defaults
 - [Enhancement] Provide a `finished?` method to the jobs for advanced reference based job schedulers.
 - [Enhancement] Provide `client.reset` notification event.